backend/requirements.txt
flask
flask_cors
pyjwt
psycopg2
bcrypt
flask-socketio
flask_sqlalchemy